Implemented authentication and authorization in the diaper and feeding controller.

+++ Controllers/DiaperController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity;
 using BabyTracker.Models;
 using BabyTracker.Models.ViewModels;
 using BabyTracker.Models.RepositoryModels;
@@ -15,13 +16,28 @@ namespace BabyTracker.Controllers
     public class DiaperController: Controller
     {
         private BabyTrackerContext context;
-        public DiaperController(BabyTrackerContext ctx)
+        private UserManager<IdentityUser> userManager;
+        private bool IsLoggedIn() => User.Identity.IsAuthenticated;
+        private bool IsInfantOwner(Infant infant) => infant.UserId == userManager.GetUserId(User);
+        private bool IsDiaperOwner(Diaper diaper) => diaper.Infant.UserId == userManager.GetUserId(User);
+        public DiaperController(BabyTrackerContext ctx, UserManager<IdentityUser> usrMgr)
         {
             context = ctx;
+            userManager = usrMgr;
         }
 
         public IActionResult Index(long id)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+            Infant infant = context.Infants.FirstOrDefault(i => i.InfantId == id);
+            if (!IsInfantOwner(infant))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+
             ViewData["InfantName"] =  context.Infants.FirstOrDefault(i => i.InfantId == id).FirstName;
             ViewBag.Id = id;
             IEnumerable<Diaper> Diapers = context.Diapers.Where(d => d.InfantId == id).Select(d => d);
@@ -30,7 +46,18 @@ namespace BabyTracker.Controllers
 
         public async Task<IActionResult> Details (long id)
         {
-            Diaper diaper = await context.Diapers.Include(d => d.Infant).FirstOrDefaultAsync(d => d.DiaperId == id);
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+
+            Diaper diaper = await context.Diapers.Include(f => f.Infant).FirstOrDefaultAsync(f => f.DiaperId == id);
+    
+            if (!IsDiaperOwner(diaper))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+            
             Infant infant = diaper.Infant;
             DiaperViewModel model = DiaperViewModelFactory.Details(diaper, infant);
             return View("DiaperEditor", model);
@@ -39,7 +66,16 @@ namespace BabyTracker.Controllers
         // HTTP Get
         public IActionResult Create (long id)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
             Infant infant = context.Infants.FirstOrDefault(i => i.InfantId == id);
+            if (!IsInfantOwner(infant))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+
             Diaper diaper = new Diaper
             {
                 Time=DateTime.Now,
@@ -51,6 +87,17 @@ namespace BabyTracker.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromForm] Diaper diaper)
         {
+
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+    
+            if (!IsInfantOwner(diaper.Infant))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+            
             if (ModelState.IsValid)
             {
                 diaper.DiaperId = default;
@@ -65,13 +112,34 @@ namespace BabyTracker.Controllers
         // HTTP Get
         public async Task<IActionResult> Edit (long id)
         {
-            Diaper diaper = await context.Diapers.Include(d => d.Infant).FirstOrDefaultAsync(d => d.DiaperId == id);
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+
+            Diaper diaper = await context.Diapers.Include(f => f.Infant).FirstOrDefaultAsync(f => f.DiaperId == id);
+            
+            if (!IsDiaperOwner(diaper))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+
             return View("DiaperEditor", DiaperViewModelFactory.Edit(diaper, diaper.Infant));
         }
 
         [HttpPost]
         public async Task<IActionResult> Edit([FromForm] Diaper diaper)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+    
+            if (!IsDiaperOwner(diaper))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+
             if (ModelState.IsValid)
             {
                 context.Diapers.Update(diaper);
@@ -84,13 +152,34 @@ namespace BabyTracker.Controllers
         // HTTP GEt
         public async Task<IActionResult> Delete (long id)
         {
-            Diaper diaper = await context.Diapers.Include(d => d.Infant).FirstOrDefaultAsync(d => d.DiaperId == id);
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+
+            Diaper diaper = await context.Diapers.Include(f => f.Infant).FirstOrDefaultAsync(f => f.DiaperId == id);
+            
+            if (!IsDiaperOwner(diaper))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+
             return View("DiaperEditor", DiaperViewModelFactory.Delete(diaper, diaper.Infant));
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(Diaper diaper)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Error/Unauthenticated");
+            }
+    
+            if (!IsDiaperOwner(diaper))
+            {
+                return RedirectToPage("/Error/Unauthorized");
+            }
+            
             long infantId = diaper.InfantId;
             context.Diapers.Remove(diaper);
             await context.SaveChangesAsync();
